+++ CHANGES.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.13.0 - unreleased
+### Changed
+* ```Renderer::GetClipRect``` now returns ```Optional<Rect>``` instead of (possibly empty) ```Rect```
+
 ### Removed
 * Deprecated ```Renderer::GetInfo()``` variant which takes pointer (use variant which takes reference)